fix(Balance): Show virtual groups
