+++ src/selectors/index.js
@@ -1,6 +1,5 @@
 import { createSelector } from 'reselect'
-import { groupBy } from 'lodash'
-import { GROUP_DOCTYPE } from 'doctypes'
+import { buildVirtualGroups } from 'ducks/groups/helpers'
 
 const getCollection = (state, attr) => {
   const col = state[attr]
@@ -28,19 +27,7 @@ export const getBills = state => {
 }
 
 export const getVirtualGroups = createSelector([getAccounts], accounts => {
-  const accountsByType = groupBy(accounts, account => account.type)
-
-  const virtualGroups = Object.entries(accountsByType)
-    .filter(([type, accounts]) => type !== 'undefined' && accounts.length > 1)
-    .map(([type, accounts]) => ({
-      _id: type,
-      _type: GROUP_DOCTYPE,
-      label: type.toLowerCase(),
-      accounts: accounts.map(account => account._id),
-      virtual: true
-    }))
-
-  return virtualGroups
+  return buildVirtualGroups(accounts)
 })
 
 export const getAllGroups = createSelector(
